Refactor diem thi handling with shared base class
