+++ Chương_trình_quản_lý_thí_sinh_dự_thi_đại_học/(Diem)DiemKhoiC.cs
@@ -2,34 +2,17 @@ using System;
 
 namespace Chương_trình_quản_lý_thí_sinh_dự_thi_đại_học
 {
-    public class DiemKhoiC : IDiemThi
+    public class DiemKhoiC : DiemThiBase
     {
         public double Van { get; set; }
         public double Su { get; set; }
         public double Dia { get; set; }
 
-        public void NhapDiem()
+        protected override (string TenMon, Func<double> Getter, Action<double> Setter)[] MonHoc => new[]
         {
-            Van = NhapDiemMon("Văn");
-            Su = NhapDiemMon("Sử");
-            Dia = NhapDiemMon("Địa");
-        }
-
-        public void InDiem()
-        {
-            Console.WriteLine($"Văn: {Van} | Sử: {Su} | Địa: {Dia}");
-        }
-
-        private static double NhapDiemMon(string tenMon)
-        {
-            double diem;
-            Console.Write($"Nhập điểm {tenMon}: ");
-            while (!double.TryParse(Console.ReadLine(), out diem) || diem < 0 || diem > 10)
-            {
-                Console.Write($"Điểm không hợp lệ! Nhập lại điểm {tenMon} (0-10): ");
-            }
-
-            return Math.Round(diem, 2);
-        }
+            ("Văn", () => Van, value => Van = value),
+            ("Sử", () => Su, value => Su = value),
+            ("Địa", () => Dia, value => Dia = value),
+        };
     }
 }
