+++ Chương_trình_quản_lý_thí_sinh_dự_thi_đại_học/Chương_trình_quản_lý_thí_sinh_dự_thi_đại_học.csproj
@@ -47,6 +47,7 @@
     <Compile Include="%28Diem%29DiemKhoiA.cs" />
     <Compile Include="%28Diem%29DiemKhoiB.cs" />
     <Compile Include="%28Diem%29DiemKhoiC.cs" />
+    <Compile Include="%28Diem%29DiemThiBase.cs" />
     <Compile Include="%28TS%29IThiKhoi.cs" />
     <Compile Include="%28Diem%29IDiemThi.cs" />
     <Compile Include="Program.cs" />
